+++ src/gii/default/email-template.php
@@ -54,15 +54,25 @@ class <?= $className ?> extends Migration
             ->scalar();
 
 <?php if (Yii::$app->getModule('email-templates')): ?>
-    <?php foreach(\motion\i18n\helpers\LanguageHelper::getInstance()->getLocales() as $language): ?>
-        $this->insert($this->translationTableName, [
-            'templateId'    => $templateId,
-            'language'      => '<?= $language ?>',
-            'subject'       => '<?= $generator->subject ?>',
-            'body'          => '<?= $generator->body ?>',
-            'hint'          => '<?= $generator->hint ?>',
-        ]);
-    <?php endforeach; ?>
+        $this->batchInsert(
+            $this->translationTableName,
+            [
+                'templateId',
+                'language',
+                'subject',
+                'body',
+                'hint',
+            ],
+            [
+            <?php foreach(\motion\i18n\helpers\LanguageHelper::getInstance()->getLocales() as $language): ?>    [
+                    'templateId'    => $templateId,
+                    'language'      => '<?= $language ?>',
+                    'subject'       => '<?= $generator->subject ?>',
+                    'body'          => '<?= $generator->body ?>',
+                    'hint'          => '<?= $generator->hint ?>',
+                ],
+            <?php endforeach; ?>]
+        );
 <?php else: ?>
         $this->insert($this->translationTableName, [
             'templateId'    => $templateId,